Changed from doubles to floats.
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@530 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/model/data/continuous/GenericContinuousTrack.java
@@ -81,12 +81,12 @@ public class GenericContinuousTrack extends ContinuousTrack {
 
                 savantFile.seek((i-1)*recordSize);
 
-                double sum = 0.0;
+                float sum = 0.0f;
                 int j;
                 try
                 {
                     for (j = 0; j < contiguousSamples; j++) {
-                        sum += savantFile.readDouble();
+                        sum += savantFile.readFloat();
                     }
                 } catch (Exception e) { break; }
                 int pos;
